Add UI.
Add font handling.
Added Render::DynamicText.
Added UI::Image.
Added UI::Text.
Added Render::TextCore.
Added ChangeTextCore.
Bounds now exists in world space or screen space.
Change BoundsMoved, BoundsScaled and BoundsRotated to have Arca::RelicID and their respective previous values.
Rename MoveBoundsTo to MoveBounds.
Delete MoveBoundsBy.
Delete MoveBoundsDirection.
Delete ScreenPoint and ScreenSize.
Delete RelativeBounds.
Delete RelativeImage.
Delete modulators.

Hold

+++ Atmos/FileCurator.cpp
@@ -13,7 +13,7 @@ namespace Atmos::File
     {
         Inscription::File::InputBinary file(command.filePath);
         file.SeekFromEnd(0);
-        const auto length = file.Position();
+        const auto length = static_cast<size_t>(file.Position());
         file.SeekFromBeginning(0);
 
         Buffer buffer;
